[VSPC-208] added review comments changes
