+++ vspace/src/main/java/edu/asu/diging/vspace/core/data/display/SpaceDisplayRepository.java
@@ -1,11 +1,14 @@
 package edu.asu.diging.vspace.core.data.display;
 
+import java.util.List;
+
 import org.javers.spring.annotation.JaversSpringDataAuditable;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.PagingAndSortingRepository;
 import org.springframework.stereotype.Repository;
 
+import edu.asu.diging.vspace.core.model.ISpace;
 import edu.asu.diging.vspace.core.model.display.impl.SpaceDisplay;
 
 @Repository
@@ -15,5 +18,7 @@ public interface SpaceDisplayRepository extends PagingAndSortingRepository<Space
     @Modifying
     @Query("delete from SpaceDisplay where space_id = ?1")
     void deleteBySpaceId(String id);
+    
+    public List<SpaceDisplay> getBySpace(ISpace space);
     
 }
