+++ vspace/src/main/java/edu/asu/diging/vspace/core/services/ISpaceManager.java
@@ -38,6 +38,4 @@ public interface ISpaceManager {
     Iterable<Space> addIncomingLinkInfoToSpaces(Iterable<Space> spaces);
     
     Page<ISpace> findByNameOrDescription(Pageable requestedPage,String searchText);
-
-    String getSpaceId(String id);
 }
